refactpr(step3): 역 생성 TestFixture 클래스 추가
- utils 패키지 경로 추가
- Line 관련 테스트마다 반복되는 역 생성 로직 분리
