+++ README.md
@@ -132,5 +132,6 @@
 - [x] Section 컬렉션에 대해서 일급 컬렉션으로 리팩토링 해보기
 - [x] 노선 삭제 인수 테스트 실패 이슈
   - Section - Line 간 FK 설정, Line 삭제 시 Section이 함께 삭제되도록 설정
-- [ ] 테스트별로 반복되는 코드들을 fixture 로 관리해보기
+- [x] 테스트별로 반복되는 코드들을 fixture 로 관리해보기
+  - 테스트마다 사용되는 역 생성 부분을 StationFixture 클래스로 분리, 필요한 갯수만큼 생성 후 ID값을 리턴받도록 하는 함수 추가
 - Section / Line 도메인 구조 수정
