+++ src/test/java/subway/StationAcceptanceTest.java
@@ -14,7 +14,7 @@ import java.util.List;
 import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static subway.AcceptanceMethods.makeStation;
+import static subway.utils.AcceptanceMethods.makeStation;
 
 @DisplayName("지하철역 관련 기능")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
